feat(ui): TE-1206 moved network call cancellation logic into a wrapper and exposed it via react context
