+++ thirdeye-ui/src/app/app.tsx
@@ -12,16 +12,10 @@
  * See the License for the specific language governing permissions and limitations under
  * the License.
  */
-import axios, { CancelTokenSource } from "axios";
-import React, {
-    FunctionComponent,
-    useEffect,
-    useLayoutEffect,
-    useState,
-} from "react";
+import React, { FunctionComponent, useEffect } from "react";
 import { useTranslation } from "react-i18next";
-import { useLocation } from "react-router-dom";
 import { AppBarConfigProvider } from "./components/app-bar/app-bar-config-provider/app-bar-config-provider.component";
+import { CancelNetworkCallsWrapper } from "./components/cancel-network-calls-wrapper/cancel-network-calls-wrapper.component";
 import {
     AppContainerV1,
     NotificationScopeV1,
@@ -41,39 +35,6 @@ export const App: FunctionComponent = () => {
     const { t } = useTranslation();
     const { notify } = useNotificationProviderV1();
 
-    // TODO: Move to a separate component
-    const [interceptorId, setInterceptorId] = useState<number | null>();
-
-    const getNewCancelToken = (): CancelTokenSource =>
-        axios.CancelToken.source();
-
-    const [cancelTokenSource, setCancelTokenSource] =
-        useState<CancelTokenSource>(getNewCancelToken());
-
-    const cancelApiCalls = (): void => {
-        cancelTokenSource.cancel("cancelApiCalls invoked");
-        setCancelTokenSource(getNewCancelToken());
-    };
-
-    useLayoutEffect(() => {
-        interceptorId && axios.interceptors.request.eject(interceptorId);
-
-        const newInterceptorId = axios.interceptors.request.use((config) => ({
-            ...config,
-            cancelToken: cancelTokenSource.token,
-        }));
-
-        setInterceptorId(newInterceptorId);
-    }, [cancelTokenSource]);
-
-    const location = useLocation();
-
-    useLayoutEffect(() => {
-        return () => {
-            cancelApiCalls();
-        };
-    }, [location.pathname]);
-
     useEffect(() => {
         if (authDisabled && authDisabledNotification) {
             // Auth disabled and user not yet notified
@@ -88,10 +49,12 @@ export const App: FunctionComponent = () => {
     }, [authDisabled, authDisabledNotification]);
 
     return (
-        <AppContainerV1 name={t("label.thirdeye")}>
-            <AppBarConfigProvider>
-                <AppRouter />
-            </AppBarConfigProvider>
-        </AppContainerV1>
+        <CancelNetworkCallsWrapper>
+            <AppContainerV1 name={t("label.thirdeye")}>
+                <AppBarConfigProvider>
+                    <AppRouter />
+                </AppBarConfigProvider>
+            </AppContainerV1>
+        </CancelNetworkCallsWrapper>
     );
 };
