+++ thirdeye-ui/src/app/components/cancel-network-calls-wrapper/cancel-network-calls-wrapper.utils.ts
@@ -1,0 +1,29 @@
+/*
+ * Copyright 2022 StarTree Inc
+ *
+ * Licensed under the StarTree Community License (the "License"); you may not use
+ * this file except in compliance with the License. You may obtain a copy of the
+ * License at http://www.startree.ai/legal/startree-community-license
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the
+ * License is distributed on an "AS IS" BASIS, WITHOUT * WARRANTIES OF ANY KIND,
+ * either express or implied.
+ *
+ * See the License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
+import axios, { CancelTokenSource } from "axios";
+import { createContext, useContext } from "react";
+import type { CancelNetworkCallsContextProps } from "./cancel-network-calls-wrapper.interfaces";
+
+export const CancelNetworkCallsContext =
+    createContext<CancelNetworkCallsContextProps>(
+        {} as CancelNetworkCallsContextProps
+    );
+
+export const useCancelNetworkCallsContext =
+    (): CancelNetworkCallsContextProps => useContext(CancelNetworkCallsContext);
+
+export const getNewCancelToken = (): CancelTokenSource =>
+    axios.CancelToken.source();
